HandBrakeCLI-wrapper/requirements.txt
asn1crypto==1.0.1
certifi==2019.11.28
cffi==1.12.3
cryptography==2.7
docutils==0.15.2
idna==2.8
jmespath==0.9.4
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
python-dateutil==2.8.0
PyYAML==5.1.2
pyzmq==18.1.1
sarge==0.1.5.post0
scramp==1.1.0
six==1.12.0
urllib3==1.25.6
